requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-Mail==0.9.1
Flask-Cors==4.0.0

python-dotenv==1.0.1
PyCryptodome==3.20.0
python-magic==0.4.27
Pillow==10.2.0
twilio==9.0.1

eventlet==0.33.3
gunicorn==21.2.0